+++ packages/commercetools/api-client/src/api/createCart/index.ts
@@ -1,5 +1,5 @@
 import { CartDraft } from './../../types/GraphQL';
-import { apolloClient, getSettings } from './../../index';
+import { apolloClient, CustomQueryFn, getSettings } from './../../index';
 import { getCustomQuery } from './../../helpers/queries';
 import defaultQuery from './defaultMutation';
 import gql from 'graphql-tag';
@@ -8,7 +8,7 @@ interface CartData extends Omit<CartDraft, 'currency'> {
   currency?: string;
 }
 
-const createCart = async (cartDraft: CartData = {}, customQueryFn?) => {
+const createCart = async (cartDraft: CartData = {}, customQueryFn?: CustomQueryFn) => {
   const { locale, acceptLanguage, currency } = getSettings();
 
   const defaultVariables = {
@@ -22,9 +22,7 @@ const createCart = async (cartDraft: CartData = {}, customQueryFn?) => {
   const { query, variables } = getCustomQuery(customQueryFn, { defaultQuery, defaultVariables });
 
   const request = await apolloClient.mutate({
-    mutation: gql`
-      ${query}
-    `,
+    mutation: gql`${query}`,
     variables,
     fetchPolicy: 'no-cache'
   });