+++ packages/core/docs/commercetools/changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.2.5 - not released
 
+- `customQuery` for checkout composables ([#5025](https://github.com/DivanteLtd/vue-storefront/issues/5025))
 
 ## 0.2.4